Move formatting func to format module
